refactor: calculate에서 calculateInfixNotation로 수정
